+++ src/IA/Heuristique.java
@@ -2,5 +2,5 @@ package IA;
 import Modele.Jeu;
 
 public interface Heuristique {
-    public int evaluer(Jeu j, int pdvJoueur);
+    public double evaluer(Jeu j, int pdvJoueur);
 }
